refactor(tui): extract file-tree path-to-tree builder into filetree/tree.ts
filetree/git.ts combined git wrappers with the path-to-tree hierarchy
builder. The tree builder has no git dependency and is consumed by both
the Solid file-tree pane (via rows.ts) and the React port (FileTree.tsx).
Move TreeNode, buildTree, and sortTree into a dedicated framework-free
module so git.ts can stay focused on git I/O.

filetree/git.ts drops from 466 to 412 lines; the new module is 60 lines.
